backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiohttp==3.9.5
google-generativeai==0.4.1
openai==1.14.0
pytest==7.4.4
httpx==0.27.0
grpcio==1.62.1
protobuf==4.25.3
google-api-core==2.19.2
google-auth==2.27.0
python-multipart==0.0.6
websockets==11.0.3